Issue #121 Split JsonPath expression AST into dedicated files
