+++ json-java21-jsonpath/src/main/java/json/java21/jsonpath/JsonPathAst.java
@@ -77,54 +77,3 @@ record JsonPathAst(List<Step> steps) {
         }
     }
 }
-
-sealed interface BoolExpr permits BoolExpr.Exists, BoolExpr.Compare {
-    record Exists(ValueExpr.Path path) implements BoolExpr {
-        public Exists {
-            Objects.requireNonNull(path, "path must not be null");
-        }
-    }
-
-    enum CmpOp {
-        LT, LE, GT, GE, EQ, NE
-    }
-
-    record Compare(CmpOp op, ValueExpr left, ValueExpr right) implements BoolExpr {
-        public Compare {
-            Objects.requireNonNull(op, "op must not be null");
-            Objects.requireNonNull(left, "left must not be null");
-            Objects.requireNonNull(right, "right must not be null");
-        }
-    }
-}
-
-sealed interface ValueExpr permits ValueExpr.Path, ValueExpr.Num, ValueExpr.Str, ValueExpr.Arith {
-
-    record Path(List<String> parts) implements ValueExpr {
-        public Path {
-            Objects.requireNonNull(parts, "parts must not be null");
-            parts.forEach(p -> Objects.requireNonNull(p, "path part must not be null"));
-        }
-    }
-
-    record Num(double value) implements ValueExpr {}
-
-    record Str(String value) implements ValueExpr {
-        public Str {
-            Objects.requireNonNull(value, "value must not be null");
-        }
-    }
-
-    enum ArithOp {
-        ADD, SUB
-    }
-
-    record Arith(ArithOp op, ValueExpr left, ValueExpr right) implements ValueExpr {
-        public Arith {
-            Objects.requireNonNull(op, "op must not be null");
-            Objects.requireNonNull(left, "left must not be null");
-            Objects.requireNonNull(right, "right must not be null");
-        }
-    }
-}
-
